moved static functions from parser in its own body
