+++ src/lexer.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <initializer_list>
 #include <string>
 #include <vector>
 #include <string_view>
@@ -64,6 +65,7 @@ class lexer
     return match(type) || match(args...);
   }
   [[nodiscard]] bool check(token_type type) const noexcept;
+  [[nodiscard]] bool check(std::initializer_list<token_type> types) const noexcept;
   [[nodiscard]] bool match(token_type type) noexcept;
 
   void skip_linebreaks();
